Extract `call_rpc` crate
