+++ rust/services/call/db/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "call_db"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 alloy-primitives = { workspace = true }
@@ -11,3 +11,6 @@ mpt = { workspace = true }
 provider = { workspace = true }
 revm = { workspace = true }
 thiserror = { workspace = true }
+
+[lints]
+workspace = true
