+++ rust/services/call/host/src/host/tests/teleport.rs
@@ -5,6 +5,7 @@ use alloy_primitives::{ChainId, b256};
 use call_common::{ExecutionLocation, RevmDB};
 use call_db::ProviderDb;
 use call_engine::verifier::teleport::fetch_latest_confirmed_l2_block;
+use call_rpc::{quicknode_op_sepolia_url, rpc_urls};
 use ethers_core::types::U64;
 use jsonrpsee::http_client::HttpClientBuilder;
 use optimism::{
@@ -14,8 +15,6 @@ use optimism::{
     types::SequencerOutput,
 };
 use provider::{BlockTag, EthersProviderFactory, ProviderFactory};
-
-use crate::test_harness::rpc::{quicknode_op_sepolia_url, rpc_urls};
 
 fn get_db(location: ExecutionLocation) -> anyhow::Result<impl RevmDB> {
     let provider_factory = EthersProviderFactory::new(rpc_urls());
