+++ rust/services/call/rpc/.env.example
@@ -1,0 +1,4 @@
+# This file is an example of how your .env file should look like.
+# Create .env file in this same directory and put your real keys in it.
+ALCHEMY_KEY=your_alchemy_key_here
+QUICKNODE_KEY=your_quicknode_key_here
